@@ -1,7 +1,6 @@
 class InquiriesController < ApplicationController
-	
-	load_and_authorize_resource
-	skip_authorize_resource :only => :new
+
+	before_action :authenticate_admin!, only: [:update, :edit, :destroy]
 
 	def new
 		@inquiry = Inquiry.new
@@ -10,24 +9,10 @@ class InquiriesController < ApplicationController
 
 	def create
   		@inquiry = Inquiry.new(inquiry_params)
-  		respond_to do |format| 
 	 		if @inquiry.save
-	  			format.html { 
-	  				redirect_to root_url, 
-	  				notice: 'Your work request has been successfully sent. We will contact you in the near future!' 
-	  			}
-	  			format.json {
-	  				render json: @inquiry, 
-	  				status: created,
-	  				location: @inquiry
-	  			}
-	  		else
-	  			format.html { render action: "new" }
-	  			format.json { 
-	  				render json: @inquiry.errors, 
-	  				status: :unprocessable_entity 
-	  			}
-	  		end
+	  		redirect_to root_url, notice: 'Your work request has been successfully sent. We will contact you in the near future!'
+	  	else
+	  		render action: "new"
 	  	end
 	end
 
@@ -38,19 +23,17 @@ class InquiriesController < ApplicationController
 	def destroy
 		@inquiry = Inquiry.find(params[:id])
 		@inquiry.destroy
-		authorize! :manage, @inquiry
 		redirect_to admin_path
 	end
 
 	def edit
 		@inquiry = Inquiry.find(params[:id])
-		authorize! :manage, @inquiry
 	end
 
 	def update
 		@inquiry = Inquiry.find(params[:id])
 		if @inquiry.update(inquiry_params)
-			redirect_to @inquiry 
+			redirect_to @inquiry
 		else
 			render 'edit'
 		end
