@@ -1,17 +1,14 @@
  class CommentsController < ApplicationController
 
- load_and_authorize_resource
-
- def create
+  def create
     @post = Post.find(params[:post_id])
     @comment = @post.comments.create(comment_params)
     redirect_to @post
   end
- 
+
   def edit
     @post = Post.find(params[:id])
     @comment = @post.comments.find(comment_params)
-    authorize! :manage, @comment 
   end
 
   def update
@@ -29,7 +26,7 @@
     @comment.destroy
     redirect_to @post
   end
- 
+
   private
     def comment_params
       params.require(:comment).permit(:commenter, :body)
