@@ -1,7 +1,6 @@
 class Post < ActiveRecord::Base
 	has_many :comments, dependent: :destroy
-	validates :title, presence: true, length: { maximum: 50 }
-	validates :text, presence: true, length: { minimum: 5 }
-	mount_uploader :title_image, TitleImageUploader
+	validates :title, presence: true, length: { maximum: 150 }
+	validates :body, presence: true, length: { minimum: 5 }
 	acts_as_taggable
 end
