@@ -1,11 +1,7 @@
 Rails.application.routes.draw do
-  
-  devise_for :users
-  
-  resources :posts do 
-    resources :comments
-  end
 
+  devise_for :admins
+  resources :posts
   resources :pages
   resources :inquiries
 
