@@ -2,8 +2,9 @@ class CreatePosts < ActiveRecord::Migration
   def change
     create_table :posts do |t|
       t.string :title
-      t.text :text
-
+      t.text :body
+      t.boolean :published
+      t.string :tagline
       t.timestamps
     end
   end
