@@ -2,12 +2,19 @@ require 'faker'
 
 FactoryGirl.define do
 	factory :post do |f|
-		f.title "Testing Post"
-		f.text "This is a simple testing post. I hope it works."
-		f.img 'test.jpg'
-	end
 
-	factory :invalid_post, parent: :post do |f|
-		f.title nil
+		f.title 		{ Faker::Lorem.sentence }
+		f.tagline		{ Faker::Lorem.sentence }
+		f.body			{
+									Faker::Lorem.paragraph +
+									"\n\n" + Faker::Lorem.paragraph(10) +
+									"\n\n" + Faker::Lorem.paragraph(10) +
+									"\n\n" + Faker::Lorem.paragraph(10) +
+									"\n\n" + Faker::Lorem.paragraph(10) +
+									"\n\n" + Faker::Lorem.paragraph(10) +
+									"\n\n" + Faker::Lorem.paragraph(10) +
+									"\n\n" + Faker::Lorem.paragraph(10) 
+								}
+		f.published	{ true }
 	end
 end
